refactor(Toast): 컴포넌트 네이밍 변경 및 아이콘 맵핑 적용, 전역 폰트 중복 제거
